--- a/hw10_program_optimization/.golangci.yml
+++ b/hw10_program_optimization/.golangci.yml
@@ -1,0 +1,5 @@
+linters-settings:
+  lll:
+    # max line length, lines longer will be reported. Default is 120.
+    # '\t' is counted as 1 character by default, and can be changed with the tab-width option
+    line-length: 170