## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [ ] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [ ] Я запустил `go mod tidy`.
- [ ] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [ ] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

### Достигнутая оптимзация
``` 
name     old time/op    new time/op    delta
Heavy       725ms ± 2%     208ms ± 2%   -71.35%  (p=0.008 n=5+5)
Heavy-4     805ms ±15%     211ms ± 3%   -73.73%  (p=0.008 n=5+5)
Light       195ms ± 7%       6ms ± 3%   -97.00%  (p=0.008 n=5+5)
Light-4     203ms ± 3%       6ms ± 4%   -97.07%  (p=0.008 n=5+5)

name     old alloc/op   new alloc/op   delta
Heavy       305MB ± 0%      12MB ± 0%   -95.95%  (p=0.008 n=5+5)
Heavy-4     305MB ± 0%      12MB ± 0%   -95.95%  (p=0.008 n=5+5)
Light       133MB ± 0%       0MB ± 0%  -100.00%  (p=0.000 n=4+5)
Light-4     133MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)

name     old allocs/op  new allocs/op  delta
Heavy       3.05M ± 0%     0.70M ± 0%   -76.98%  (p=0.008 n=5+5)
Heavy-4     3.05M ± 0%     0.70M ± 0%   -76.98%  (p=0.008 n=5+5)
Light       1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
Light-4     1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
```